Pushing some old attempts to fix DLL interface support
Trying to get the game to load with DLL/SOs at least. (64 and 32)

Still doesn't work.  Need to come up with something with the Ghoul2 changes.  Because of the ways that SOF2 handles some G2 API calls and the differences in Mod API -> Renderer work now, it simply crashes.  Haven't been able to track exactly why its getting bad data but it is.  Particularly with CG_G2_GETBOLTINDEX.